requirements.txt
# Core ML/DL
torch>=2.0.0
transformers>=4.15.0
sentence-transformers>=2.2.0
scikit-learn>=1.0.2
numpy>=1.21.0
pandas>=1.4.0

# NLP & Text Processing
nltk>=3.8.1
spacy>=3.5.0
bert-score>=0.3.11
beautifulsoup4>=4.9.3
lxml>=4.9.0

# Visualization & Dashboard
plotly>=5.3.0
dash>=2.0.0
streamlit>=1.2.0
dash-bootstrap-components>=1.0.0
dash-core-components>=2.0.0
dash-html-components>=2.0.0
dash-table>=5.0.0
jupyter-dash>=0.4.0

# Clustering
hdbscan>=0.8.29
umap-learn>=0.5.3
networkx>=2.8.0
python-louvain>=0.16
scipy>=1.8.0

# Utilities & Performance
tqdm>=4.67.1
pyyaml>=6.0
python-dotenv>=0.19.0
cachetools>=5.0.0
joblib>=1.1.0
psutil>=5.9.0

# Testing & Development
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.3.0
isort>=5.10.0
flake8>=4.0.0
mypy>=0.950

# Deep Learning & Embeddings
datasets>=2.15.0

# NLP & Text Processing
tokenizers>=0.13.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Web Framework & Dashboard
flask>=2.0.0

# Documentation
sphinx>=4.3.0
sphinx-rtd-theme>=1.0.0

# Optional: GPU Support
# cudatoolkit>=11.3  # Uncomment if using GPU

# Logging & Monitoring
tensorboard>=2.7.0
wandb>=0.12.0

# File Handling
pathlib>=1.0.1
PyPDF2>=2.0.0
python-docx>=0.8.11

# API Integration
requests>=2.32.3
aiohttp>=3.8.0

# Testing
pytest-asyncio>=0.16.0

# Multi-core CPU and GPU Optimization

# New dependencies for adaptive style selection
textstat>=0.7.3  # For text complexity analysis
lexical-diversity>=0.1.1  # For lexical diversity metrics
language-tool-python>=2.7.1  # For grammar checking

# Domain-agnostic processing
textacy>=0.12.0

# Graph Clustering Dependencies
networkx>=2.8.0
python-louvain>=0.16  # Replaces community package
scipy>=1.8.0  # Required by networkx

# Add these lines to requirements.txt
beautifulsoup4>=4.9.3
lxml>=4.9.0  # Optional but recommended parser for bs4
spacy>=3.5.0

# Add missing dependencies

# Dashboard & Visualization
dash>=2.0.0
dash-bootstrap-components>=1.0.0
dash-core-components>=2.0.0
dash-html-components>=2.0.0
dash-table>=5.0.0
jupyter-dash>=0.4.0

# Performance Optimization
cachetools>=5.0.0
joblib>=1.1.0

# API and Data Processing
requests>=2.32.3
beautifulsoup4>=4.9.3
lxml>=4.9.0

# Dashboard
dash>=2.0.0
dash-bootstrap-components>=1.0.0
dash-core-components>=2.0.0
dash-html-components>=2.0.0
dash-table>=5.0.0
jupyter-dash>=0.4.0

# Visualization
plotly>=5.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-asyncio>=0.16.0

# Development
black>=22.3.0
isort>=5.10.0
flake8>=4.0.0
mypy>=0.950

# Model Training
rouge-score>=0.1.2
bert-score>=0.3.11
tabulate>=0.8.9
transformers>=4.15.0
datasets>=2.0.0
torch>=1.9.0
accelerate>=0.5.1
